Yet more Consistency & Fixes

+++ src/nvr_at.c
@@ -880,7 +880,7 @@ nvr_start(nvr_t *nvr)
     if (machines[machine].flags & MACHINE_COREBOOT) {
         /* Sync floppy drive types on coreboot machines, as SeaBIOS
            lacks a setup utility and just leaves these untouched. */
-        int fdd;
+        uint8_t fdd;
 
         nvr->regs[RTC_FDD_TYPES] = 0x00;
         nvr->regs[RTC_INST_EQUIP] |= 0xc0;
